+++ src/components/index.ts
@@ -12,7 +12,7 @@ export {InputTime, InputTimeProps} from "./input-time";
 export {Label, LabelProps, LabelStyle} from "./label";
 export {Panel, PanelButtons, PanelProps, PanelStyle} from "./panel";
 export {Popin, PopinStyle} from "./popin";
-export {Scrollable, ScrollableContext, Sticky} from "./scrollable";
+export {Scrollable, ScrollableContext, ScrollableStyle, Sticky} from "./scrollable";
 export {
     ScrollspyContainer,
     ScrollspyContainerProps,
