+++ src/layout/content.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import {useTheme} from "../theme";
+import {LayoutStyle, styles} from "./layout";
+
+export function Content(props: {children?: React.ReactNode; theme?: LayoutStyle}) {
+    const theme = useTheme("popin", styles, props.theme);
+    return <div className={theme.content}>{props.children}</div>;
+}
