+++ src/style/header.css
@@ -7,7 +7,7 @@
     --header-color: var(--color-black);
     --header-content-background: var(--color-white);
     --header-content-padding: 20px 10px;
-    --header-margin-bottom: 50px;
+    --header-margin-bottom: 40px;
     --header-toprow-background: var(--color-white);
     --header-toprow-height: 60px;
     --header-toprow-padding: 10px;
