+++ src/style/scrollspy-container.css
@@ -4,7 +4,7 @@
 :root {
     --scrollspy-active-border-color: var(--color-primary);
     --scrollspy-active-color: var(--color-accent);
-    --scrollspy-padding-top: 10px;
+    --scrollspy-padding-top: 20px;
     --scrollspy-padding-left: 40px;
     --scrollspy-padding-right: 20px;
 }
